ref(server-utils): Streamline graphql integration
